+++ src/dynamicConfig.ts
@@ -172,6 +172,7 @@ export class DynamicConfig {
 
     public ScriptFileExtensionRegex: RegExp = new RegExp('to-load msr.default.scriptFiles');
     public ConfigAndDocFilesRegex: RegExp = new RegExp('to-load msr.default.configAndDocs');
+    public ShowLongTip: boolean = true;
 
     private TmpToggleEnabledExtensionToValueMap = new Map<string, boolean>();
 
@@ -303,6 +304,7 @@ export class DynamicConfig {
         this.UseGitIgnoreFile = getConfigValue('useGitIgnoreFile') === 'true';
         this.OmitGitIgnoreExemptions = getConfigValue('omitGitIgnoreExemptions') === 'true';
         this.SkipDotFolders = getConfigValue('skipDotFoldersIfUseGitIgnoreFile') === 'true';
+        this.ShowLongTip = getConfigValue('cookCmdAlias.showLongTip') === 'true';
 
         SearchConfig.reload();